Magazine_app/requirements.txt
blinker==1.8.2
chardet==5.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
Flask==3.0.3
Flask-MySQLdb==2.0.0
fonttools==4.54.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mysqlclient==2.2.5
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
reportlab==4.2.5
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.6.0
tzdata==2024.2
Werkzeug==3.0.3
fastapi==0.115.5
uvicorn==0.32.0
requests==2.32.3
beautifulsoup4==4.12.3
fpdf2==2.7.9
pydantic==2.9.2
python-multipart==0.0.20
celery==5.3.6
redis==5.0.1

# Agent dependencies
langgraph==0.2.55
langchain-google-genai==2.0.5
langchain-community==0.3.7

# Database and auth dependencies
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
bcrypt==3.2.2